% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_count_and_prop.R
\name{group_count_and_prop}
\alias{group_count_and_prop}
\title{Count and calculate proportions in group(s)}
\usage{
group_count_and_prop(.data, ..., cond)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{...}{Variables to group by.}

\item{cond}{Short for "condition". Expressions that return a logical value. Expressions are defined in terms of variables in .data and multiple expressions, if included, must be combined with the & operator. Only rows for which all conditions return TRUE are evaluated.}
}
\value{
A new data frame summarizing the groups, total number of observations in each group, count of observations that return TRUE for the condition(s) and proportion (\%) of observations that return TRUE for the condition(s)
}
\description{
Summarizes the number and proportion of observations that meet one or more conditions within given group(s).
}
\examples{
group_count_and_prop(dplyr::starwars, gender, cond = height > 150 & mass > 50)
group_count_and_prop(mtcars, cyl, cond = mpg > 20)
}
